plan: unwrap MPI communicator on plan description, move wrapper to plan
The plan description should just have whatever the user gave us, and we have
no responsibility to free it.  Once the plan is created, duplicate the
user's communicator so we can set our own error policy on it.

Rename the description's communicator so it's clear the user owns it.

This ensures descriptions are are copyable since they're "just" data and
minimizes the number of communicator duplications that happen when we
create plans.